util.dnslookup: removed
dnslookup uses function splitport, which is not part of the official API not in
Py3 nor in Py2 (and it is not a part of six) [1].

Split port number from IPv4 AND IPv6 URLs needs more efforts. ATM I see no usage
of the dnslookup function, so I removed it for the first.

[1] https://bugs.python.org/issue27485

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>